+++ src/KGzocha/Searcher/FilterImposer/Doctrine/AbstractODMBuilderFilterImposer.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace KGzocha\Searcher\FilterImposer;
-use KGzocha\Searcher\Context\ODMBuilderSearchingContext;
+namespace KGzocha\Searcher\FilterImposer\Doctrine;
+
+use KGzocha\Searcher\Context\Doctrine\ODMBuilderSearchingContext;
 use KGzocha\Searcher\Context\SearchingContextInterface;
+use KGzocha\Searcher\FilterImposer\FilterImposerInterface;
 
 /**
  * @author Krzysztof Gzocha <krzysztof@propertyfinder.ae>
